Switch clasp logs from using google-cloud library to googleapis library (#192)
This gets rid of `@google-cloud/logging` in favor of `googleapis`.

There were not many changes to be made, though the response format is a bit different, so parsing the logs had to be changed. Additionally, new scopes are needed for reading the logs. Finally, `@google-cloud/logging` and `grpc` are removed.

Thanks to @grant for finding a simple solution to the `grpc` problem!
 
Signed-off-by: campionfellin <campionfellin@gmail.com>

Fixes #190 

- [ ] `npm run test` succeeds. (no tests really for `clasp logs`, I manually tested and it seems to work fine.
- [x] `npm run lint` succeeds.
- [ ] Appropriate changes to README are included in PR.